Load downscaled TraCE21ka files by years.

Load downscaled TraCE21ka files by years.

Usage

read_dsclim(
  folder,
  var,
  y_start,
  y_end,
  rcp = NULL,
  gcm = NULL,
  calendar_dates = FALSE,
  sf = NULL,
  proxy = TRUE
)

Arguments

folder

A character string with the path to the folder where the downscaled TraCE21ka files, ordered in folders by variables.

var

A character string with the name of the variables to be loaded.

y_start

A number with the first year to be loaded.

y_end

A number with the last year to be loaded.

rcp

TBW

gcm

TBW

calendar_dates

A logical value indicating whether dates should be corrected to calendar dates.

sf

A sf object to be used for cropping the object.

proxy

A logical value indicating whether the data should be loaded as a stars_proxy object (TRUE) or as a stars object (FALSE).

Value

A stars object with the downscaled TraCE21ka data, cropped using the sf object if provided.

Details

Years should be in the format calibrated Before Present (e.g. 0 for loading year 1950 in the gregorian calendar).

Examples

# TBW